build: edits

+++ src/Depositer.sol
@@ -61,9 +61,10 @@ contract Depositer {
 
     event Cloned(address indexed clone);
 
-    function cloneDepositer(
-        address _comet
-    ) external returns (address newDepositer) {
+    function cloneDepositer(address _comet)
+        external
+        returns (address newDepositer)
+    {
         require(original, "!original");
         newDepositer = _clone(_comet);
     }
@@ -202,9 +203,11 @@ contract Depositer {
         return accrued > claimed ? accrued - claimed : 0;
     }
 
-    function getNetBorrowApr(
-        uint256 newAmount
-    ) public view returns (uint256 netApr) {
+    function getNetBorrowApr(uint256 newAmount)
+        public
+        view
+        returns (uint256 netApr)
+    {
         uint256 newUtilization = ((comet.totalBorrow() + newAmount) * 1e18) /
             (comet.totalSupply() + newAmount);
         uint256 borrowApr = getBorrowApr(newUtilization);
@@ -216,7 +219,11 @@ contract Depositer {
     /*
      * Get the current supply APR in Compound III
      */
-    function getSupplyApr(uint256 newUtilization) public view returns (uint) {
+    function getSupplyApr(uint256 newUtilization)
+        public
+        view
+        returns (uint256)
+    {
         unchecked {
             return
                 comet.getSupplyRate(
@@ -228,9 +235,11 @@ contract Depositer {
     /*
      * Get the current borrow APR in Compound III
      */
-    function getBorrowApr(
-        uint256 newUtilization
-    ) public view returns (uint256) {
+    function getBorrowApr(uint256 newUtilization)
+        public
+        view
+        returns (uint256)
+    {
         unchecked {
             return
                 comet.getBorrowRate(
@@ -252,9 +261,11 @@ contract Depositer {
      * @param newAmount The new amount we will be supplying
      * @return The reward APR in USD as a decimal scaled up by 1e18
      */
-    function getRewardAprForSupplyBase(
-        uint256 newAmount
-    ) public view returns (uint) {
+    function getRewardAprForSupplyBase(uint256 newAmount)
+        public
+        view
+        returns (uint256)
+    {
         Comet _comet = comet;
         unchecked {
             uint256 rewardToSuppliersPerDay = _comet.baseTrackingSupplySpeed() *
@@ -274,9 +285,11 @@ contract Depositer {
      * @param newAmount The new amount we will be borrowing
      * @return The reward APR in USD as a decimal scaled up by 1e18
      */
-    function getRewardAprForBorrowBase(
-        uint256 newAmount
-    ) public view returns (uint256) {
+    function getRewardAprForBorrowBase(uint256 newAmount)
+        public
+        view
+        returns (uint256)
+    {
         // borrowBaseRewardApr = (rewardTokenPriceInUsd * rewardToBorrowersPerDay / (baseTokenTotalBorrow * baseTokenPriceInUsd)) * DAYS_PER_YEAR;
         Comet _comet = comet;
         unchecked {
